adding model utils module for utility functions
