+++ src/cabinetry/__init__.py
@@ -3,6 +3,7 @@ from . import fit  # NOQA
 from . import route  # NOQA
 from . import template_builder  # NOQA
 from . import template_postprocessor  # NOQA
+from . import model_utils  # NOQA
 from . import visualize  # NOQA
 from . import workspace  # NOQA
