+++ src/cabinetry/experimental.py
@@ -1,64 +1,16 @@
 import logging
 import pathlib
-from typing import Any, Dict, List, Tuple, Union
+from typing import Any, Dict, Union
 
 import awkward1 as ak
 import numpy as np
 import pyhf
 
 from . import configuration
+from . import model_utils
 from . import template_builder
 
 log = logging.getLogger(__name__)
-
-
-def get_asimov_parameters(model: pyhf.pdf.Model) -> Tuple[List[float], List[float]]:
-    """Returns a list of Asimov parameter values and pre-fit uncertainties for a model.
-
-    Args:
-        model (pyhf.pdf.Model): model for which to extract the parameters
-
-    Returns:
-        Tuple[List[float], List[float]]:
-            - the Asimov parameters, in the same order as
-              ``model.config.suggested_init()``
-            - pre-fit uncertainties for the parameters
-    """
-    # create a list of parameter names, one entry per single parameter
-    # (vectors like staterror expanded)
-    auxdata_pars_all = []
-    for parameter in model.config.auxdata_order:
-        auxdata_pars_all += [parameter] * model.config.param_set(parameter).n_parameters
-
-    # create a list of Asimov parameters (constrained parameters at the
-    # best-fit value from the aux measurement, unconstrained parameters at
-    # the init specified in the workspace)
-    asimov_parameters = []
-    pre_fit_unc = []  # pre-fit uncertainties for parameters
-    for parameter in model.config.par_order:
-        # indices in auxdata list that match the current parameter
-        aux_indices = [i for i, par in enumerate(auxdata_pars_all) if par == parameter]
-        if aux_indices:
-            # pick up best-fit value from auxdata
-            inits = [
-                aux for i, aux in enumerate(model.config.auxdata) if i in aux_indices
-            ]
-        else:
-            # pick up suggested inits (for normfactors)
-            inits = model.config.param_set(parameter).suggested_init
-        asimov_parameters += inits
-
-        # for constrained parameters, obtain their pre-fit uncertainty
-        if model.config.param_set(parameter).constrained:
-            pre_fit_unc += model.config.param_set(parameter).width()
-        else:
-            if model.config.param_set(parameter).n_parameters != 1:
-                # currently only supporting parameters with a single value
-                log.error("cannot handle paramater {parameter}")
-            # unconstrained parameter, do not add any uncertainties
-            pre_fit_unc.append(0.0)
-
-    return asimov_parameters, pre_fit_unc
 
 
 def calculate_stdev(
@@ -184,7 +136,7 @@ def data_MC(
         }
     )
 
-    asimov_pars, pre_fit_unc = get_asimov_parameters(model)
+    asimov_pars, pre_fit_unc = model_utils.get_asimov_parameters(model)
 
     if prefit:
         # override post-fit quantities by pre-fit settings
